refactor prep for target data
